AVR: Fix "--relax and -r may not be used together" from ld
when compiling with: -g -mrelax -flto.

lto-wrapper adds all target options to the command line of the
partial link with -r, but -r and -mrelax don't work together
as the linker throws an error when it sees -r and --relax.
Hence don't copy -mrelax to the partial link command options.

	* lto-wrapper (run_gcc): Don't add -mrelax when -r is present.